+++ src/http/http-client.ts
@@ -33,15 +33,17 @@ export class HttpClient {
 	 * @param data Data to send as query
 	 * @param bearer Bearer token to send
 	 * @param expect Array of status codes to expect
+	 * @param customOptions Additional request options
 	 * @return Returns a promise of HttpClientResponse
 	 */
 	public post(
 		path: string,
 		data: object,
 		bearer: boolean | string = false,
-		expect: number[] = [ 200, 201, 202, 204 ]
+		expect: number[] = [ 200, 201, 202, 204 ],
+		customOptions?: object
 	): Promise<HttpClientResponse> {
-		const options = {
+		let options = {
 			method: 'POST',
 			url: `${this.url}:${this.port}${path}`,
 			body: data,
@@ -54,6 +56,10 @@ export class HttpClient {
 		
 		if (this.headers) {
 			options['headers'] = this.headers;
+		}
+
+		if (customOptions) {
+			options = Object.assign(options, options);
 		}
 
 		return new Promise<HttpClientResponse>((accept, reject) => {
@@ -82,16 +88,18 @@ export class HttpClient {
 	 * @param data Data to send as query
 	 * @param bearer Bearer token to send
 	 * @param expect Array of status codes to expect
+	 * @param customOptions Additional request options
 	 * @return Returns a promise of HttpClientResponse
 	 */
 	public get(
 		path: string,
 		data: boolean | Object = false,
 		bearer: boolean | string = false,
-		expect: number[] = [ 200, 202 ]
+		expect: number[] = [ 200, 202 ],
+		customOptions?: object
 	): Promise<HttpClientResponse> {
 		return new Promise<HttpClientResponse>((accept, reject) => {
-			const options = {
+			let options = {
 				method: 'GET',
 				url: `${this.url}:${this.port}${path}`,
 				json: true,
@@ -104,6 +112,10 @@ export class HttpClient {
 		
 			if (this.headers) {
 				options['headers'] = this.headers;
+			}
+
+			if (customOptions) {
+				options = Object.assign(options, options);
 			}
 
 			this.request(options, (error, response, body) => {
@@ -131,15 +143,17 @@ export class HttpClient {
 	 * @param data Data to send as query
 	 * @param bearer Bearer token to send
 	 * @param expect Array of status codes to expect
+	 * @param customOptions Additional request options
 	 * @return Returns a promise of HttpClientResponse
 	 */
 	public patch(
 		path: string,
 		data: object,
 		bearer: boolean | string = false,
-		expect: number[] = [ 200, 201, 202, 204 ]
+		expect: number[] = [ 200, 201, 202, 204 ],
+		customOptions?: object
 	): Promise<HttpClientResponse> {
-		const options = {
+		let options = {
 			method: 'PATCH',
 			url: `${this.url}:${this.port}${path}`,
 			body: data,
@@ -152,6 +166,10 @@ export class HttpClient {
 		
 		if (this.headers) {
 			options['headers'] = this.headers;
+		}
+
+		if (customOptions) {
+			options = Object.assign(options, options);
 		}
 
 		return new Promise<HttpClientResponse>((accept, reject) => {
@@ -180,15 +198,17 @@ export class HttpClient {
 	 * @param data Data to send as query
 	 * @param bearer Bearer token to send
 	 * @param expect Array of status codes to expect
+	 * @param customOptions Additional request options
 	 * @return Returns a promise of HttpClientResponse
 	 */
 	public delete(
 		path: string,
 		bearer: boolean | string = false,
-		expect: number[] = [ 200, 202, 204 ]
+		expect: number[] = [ 200, 202, 204 ],
+		customOptions?: object
 	): Promise<HttpClientResponse> {
 		return new Promise<HttpClientResponse>((accept, reject) => {
-			const options = {
+			let options = {
 				method: 'DELETE',
 				url: `${this.url}:${this.port}${path}`
 			};
@@ -199,6 +219,10 @@ export class HttpClient {
 		
 			if (this.headers) {
 				options['headers'] = this.headers;
+			}
+
+			if (customOptions) {
+				options = Object.assign(options, options);
 			}
 
 			this.request(options, (error, response, body) => {